# docs/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "StudyAss 学习助手 API",
    "description": "StudyAss AI 学习平台 - RESTful API 接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "俊哥的学习助手团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.studyass.com/api",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/ai/generate-questions": {
      "post": {
        "tags": ["AI 出题"],
        "summary": "AI 出题",
        "operationId": "generateQuestions",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuestionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "题目生成成功"
          }
        }
      }
    },
    "/ai/task-logs": {
      "get": {
        "tags": ["AI 出题"],
        "summary": "获取任务日志列表",
        "operationId": "getTaskLogs",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "任务日志列表"
          }
        }
      }
    },
...
    "/practice/sessions": {
      "post": {
        "tags": ["练习会话"],
        "summary": "创建练习会话",
        "operationId": "createSession",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "会话创建成功"
          }
        }
      },
      "get": {
        "tags": ["练习会话"],
        "summary": "获取会话列表",
        "operationId": "getSessions",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "会话列表"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GenerateQuestionsRequest": {
        "type": "object",
        "properties": {
          "knowledge_point_id": {"type": "integer"},
          "question_type": {"type": "string", "enum": ["choice", "fill", "calculation"]},
          "difficulty": {"type": "string", "enum": ["easy", "medium", "hard"]},
          "count": {"type": "integer", "minimum": 1, "maximum": 20}
        }
      },
...
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "subject": {"type": "string"},
          "knowledge_point_ids": {"type": "array", "items": {"type": "integer"}},
          "question_count": {"type": "integer"},
          "difficulty": {"type": "string", "enum": ["easy", "medium", "hard"]},
          "time_limit": {"type": "integer"}
        }
      },
...
    }
  }
}